requirements.txt
pyttsx3>=2.90
SpeechRecognition>=3.10.0
Pillow>=8.0.0
pygame>=2.1.0
wolframalpha>=4.0.0
pyautogui>=0.9.50
matplotlib>=3.5.0
handwrite>=0.1.0